% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagegename-data.R
\docType{data}
\name{hk}
\alias{hk}
\title{Hong Kong COVID-19 Cases Dataset}
\format{
A data frame with 3 variables:
\describe{
  \item{lon}{Longitude of the location}
  \item{lat}{Latitude of the location}
  \item{t}{Number of COVID-19 cases}
}
}
\usage{
hk
}
\description{
This dataset contains the COVID-19 cases data in Hong Kong.
}
\keyword{datasets}
