% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdv.R
\name{kdv}
\alias{kdv}
\title{Use KDV}
\usage{
kdv(
  longitude,
  latitude,
  bandwidth_s = 1000,
  row_pixels = 800,
  col_pixels = 640
)
}
\arguments{
\item{longitude}{features' longitude}

\item{latitude}{features' latitude}

\item{bandwidth_s}{spatial bandwidth}

\item{row_pixels}{row pixels}

\item{col_pixels}{col pixels}
}
\value{
kdv result
}
\description{
Efficient and accurate kernel density visualization.
}
\examples{
\donttest{
data(hk)
resKDV <- kdv(hk$lon, hk$lat, 1000, 800 ,640)
}
}
