% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kdvCpp}
\alias{kdvCpp}
\title{kernel density visualization in C++}
\usage{
kdvCpp(args)
}
\arguments{
\item{args}{arguments for kdv}
}
\value{
the kdv result
}
\description{
kernel density visualization in C++
}
\keyword{internal}
