% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdv.R
\name{plotSTKDV}
\alias{plotSTKDV}
\title{Plot STKDV}
\usage{
plotSTKDV(data)
}
\arguments{
\item{data}{result of stkdv}
}
\value{
No return value, called to plot STKDV heatmap
}
\description{
Plot STKDV
}
\examples{
\donttest{
data(hk)
resSTKDV <- stkdv(hk$lon, hk$lat, hk$t, 1000, 6, 800, 640, 32)
plotSTKDV(resSTKDV)
}
}
