% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdv.R
\name{plotKDV}
\alias{plotKDV}
\title{Plot KDV}
\usage{
plotKDV(data)
}
\arguments{
\item{data}{result of kdv}
}
\description{
Plot KDV
}
\examples{
data(hk)
resKDV <-kdv(hk$lon, hk$lat, 1000.0, 800, 640)
plotKDV(resKDV)
}
